Move GitHub and GitLab api classes to ext package
