+++ tests/unit_tests/test_command.py
@@ -5,7 +5,7 @@ from unittest.mock import patch, MagicMock, call, ANY, PropertyMock
 import pytest
 
 import repobee
-import repobee.github_api
+import repobee.ext.github_api
 from repobee import command
 from repobee import git
 from repobee import util
@@ -136,8 +136,8 @@ def api_mock(request, mocker):
     def url_from_repo_info(repo_info):
         return generate_repo_url(repo_info.name)
 
-    mock = MagicMock(spec=repobee.github_api.GitHubAPI)
-    api_class = mocker.patch("repobee.github_api.GitHubAPI", autospec=True)
+    mock = MagicMock(spec=repobee.ext.github_api.GitHubAPI)
+    api_class = mocker.patch("repobee.ext.github_api.GitHubAPI", autospec=True)
     api_class.return_value = mock
 
     mock.get_repo_urls.side_effect = partial(
@@ -147,7 +147,8 @@ def api_mock(request, mocker):
         map(url_from_repo_info, repo_infos)
     )
     mock.get_issues = MagicMock(
-        spec="repobee.github_api.GitHubAPI.get_issues", side_effect=_get_issues
+        spec="repobee.ext.github_api.GitHubAPI.get_issues",
+        side_effect=_get_issues,
     )
     type(mock).token = PropertyMock(return_value=TOKEN)
     return mock
